requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
dotenv==0.9.9
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.47.0
google-genai==1.59.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langsmith==0.6.4
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.1
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
propcache==0.4.1
psycopg==3.3.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
tenacity==9.1.2
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
undetected-chromedriver==3.5.5
urllib3==2.6.3
uuid_utils==0.13.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
